requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.37